+++ client/entrypoints/amazon-pay-settings/amazon-pay-settings-section.js
@@ -3,7 +3,7 @@ import React, { useMemo } from 'react';
 import interpolateComponents from 'interpolate-components';
 import { loadStripe } from '@stripe/stripe-js';
 import styled from '@emotion/styled';
-import ExpressCheckoutPreviewComponent from './express-checkout-button-preview';
+import ExpressCheckoutPreviewComponent from './express-checkout-preview-component';
 import {
 	Card,
 	RadioControl,
